requirements.txt
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
blis==0.4.1
bs4==0.0.1
catalogue==1.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
cymem==2.0.3
decorator==4.4.1
idna==2.9
importlib-metadata==1.5.0
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
jsonschema==3.2.0
more-itertools==8.2.0
murmurhash==1.0.0
mysqlclient==1.4.6
nltk==3.4.5
numpy==1.18.1
packaging==20.1
pandas==1.0.1
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
plac==0.9.6
pluggy==0.13.0
preshed==3.0.2
prompt-toolkit==3.0.3
ptyprocess==0.6.0
py==1.8.1
pycparser==2.19
Pygments==2.5.2
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.15.7
PySocks==1.7.1
pytest==5.3.5
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
spacy==2.2.3
SQLAlchemy==1.3.13
thinc==7.3.0
tqdm==4.43.0
traitlets==4.3.3
urllib3==1.25.7
wasabi==0.6.0
wcwidth==0.1.8
zipp==3.0.0
